refactor: drop src layout
