+++ .gitignore
@@ -4,4 +4,4 @@ website/core.db
 build/
 __pycache__/
 *.sqlite3
-env/
+env
