--- app/__init__.py
@@ -1,3 +1,0 @@
-from sample_registry import __version__ as version
-
-__version__ = version
