+++ pyproject.toml
@@ -28,6 +28,8 @@ classifiers = [
 ]
 
 dependencies = [
+    "flask",
+    "flask-sqlalchemy",
     "sqlalchemy",
     "seqbackup @ git+https://github.com/PennChopMicrobiomeProgram/seqBackup.git@master",
 ]
@@ -56,7 +58,12 @@ export_samples = "sample_registry.export:export_samples"
 create_test_db = "sample_registry.db:create_test_db"
 sample_registry_version = "sample_registry:sample_registry_version"
 
-[tool.setuptools]
+[tool.setuptools.package-data]
+"sample_registry" = [
+    "templates/*.html",
+    "static/*",
+    "static/img/*",
+]
 
 [build-system]
 requires = ["setuptools>=61.0.0", "wheel"]
