+++ sample_registry/mapping.py
