+++ sample_registry/register.py
