+++ sample_registry/registrar.py
