+++ sample_registry/static/favicon.ico
