+++ sample_registry/static/img/forward_enabled.jpg
