+++ sample_registry/static/img/mainlogo.jpg
