+++ sample_registry/static/img/ricks_dna.png
