+++ sample_registry/static/img/sort_asc.png
