+++ sample_registry/static/img/sort_desc.png
