+++ sample_registry/templates/base.html
