+++ sample_registry/templates/browse_runs.html
