+++ sample_registry/templates/browse_tags.html
