+++ sample_registry/templates/show_stats.html
