+++ sample_registry/templates/show_tag.html
